Image Picker and extracted views
