+++ Tudor Bank/Tudor Bank/ImagePicker.swift
@@ -1,0 +1,51 @@
+//
+//  ImagePicker.swift
+//  Tudor Bank
+//
+//  Created by David Tudor on 15/02/2024.
+//
+
+import Foundation
+import UIKit
+import SwiftUI
+
+struct ImagePicker: UIViewControllerRepresentable {
+    @Binding var selectedImage: UIImage?
+    @Binding var isPickerShowing: Bool
+
+    func makeUIViewController(context: Context) -> some UIViewController {
+        let imagePicker = UIImagePickerController()
+        imagePicker.sourceType = .photoLibrary
+        imagePicker.delegate = context.coordinator
+        
+        return imagePicker
+    }
+    
+    func updateUIViewController(_ uiViewController: UIViewControllerType, context: Context) {
+    }
+    
+    func makeCoordinator() -> Coordinator {
+        return Coordinator(self)
+    }
+}
+
+class Coordinator: NSObject, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
+    var parent: ImagePicker
+    
+    init(_ picker: ImagePicker) {
+        self.parent = picker
+    }
+    
+    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
+        if let image = info[UIImagePickerController.InfoKey.originalImage] as? UIImage {
+            DispatchQueue.main.async {
+                self.parent.selectedImage = image
+            }
+        }
+        parent.isPickerShowing = false
+    }
+    
+    func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+        parent.isPickerShowing = false
+    }
+}
